Removed two warnings

+++ Node.cpp
@@ -130,7 +130,7 @@ void Node::insert(Node* newChild, bool b) {
 }
 
 void Node::insertChildren(vector<Symbol*> children) {
-    for(int i = 0; i < children.size(); i++) {
+    for(size_t i = 0; i < children.size(); i++) {
         insert(new Node(children[i], this, i), true);
     }
 }
